Moved debug_msg() from init.py to adafruit_wiznet5k_debug.py.
